--- src/aioway/_exprs/__init__.py
@@ -1,4 +1,0 @@
-# Copyright (c) AIoWay Authors - All Rights Reserved
-
-from .exprs import *
-from .ops import *
