+++ src/aioway/_ops/signs.py
@@ -9,8 +9,8 @@ from typing import Self
 from lark import Lark
 
 from aioway import _logging
+from aioway._exprs import Expr
 
-from ..exprs import Expr
 from . import _common
 from .types import ParamListTransformer, TypeList
